Move eth2/events.py -> trinity/sync/beacon/events.py
